Extract Popen with finalizer as test fixture (DRY)
